+++ examples/budget_project/main.tf
@@ -36,20 +36,32 @@ resource "random_string" "suffix" {
   upper   = false
 }
 
-module "app-engine-project" {
+module "budget_project" {
   source            = "../../"
-  name              = "appeng-${random_string.suffix.result}"
+  name              = "budget-project-${random_string.suffix.result}"
   random_project_id = true
   org_id            = var.org_id
   folder_id         = var.folder_id
   billing_account   = var.billing_account
-  activate_apis = [
-    "appengine.googleapis.com",
-  ]
+  budget_amount     = var.budget_amount
 }
 
-module "app-engine" {
-  source      = "../../modules/app_engine"
-  project_id  = module.app-engine-project.project_id
-  location_id = "us-east4"
+
+# An additional budget with more options
+resource "google_pubsub_topic" "budget" {
+  name    = "budget-topic-${random_string.suffix.result}"
+  project = module.budget_project.project_id
+}
+
+module "additional_budget" {
+  source = "../../modules/budget"
+
+  billing_account        = var.billing_account
+  projects               = [var.parent_project_id, module.budget_project.project_id]
+  amount                 = var.budget_amount
+  display_name           = "CI/CD Budget for ${module.budget_project.project_id}"
+  credit_types_treatment = var.budget_credit_types_treatment
+  services               = var.budget_services
+  alert_spent_percents   = var.budget_alert_spent_percents
+  alert_pubsub_topic     = "projects/${module.budget_project.project_id}/topics/${google_pubsub_topic.budget.name}"
 }