+++ modules/gsuite_enabled/variables.tf
@@ -17,7 +17,7 @@
 // This file was automatically generated from a template in ./autogen
 
 variable "random_project_id" {
-  description = "Enables project random id generation. Mutually exclusive with project_id being non-empty."
+  description = "Adds a suffix of 4 random characters to the `project_id`"
   default     = "false"
   type        = string
 }
@@ -39,8 +39,7 @@ variable "name" {
 }
 
 variable "project_id" {
-  description = "If provided, the project uses the given project ID. Mutually exclusive with random_project_id being true."
-  type        = string
+  description = "The ID to give the project. If not provided, the `name` will be used."
   default     = ""
 }
 
@@ -65,6 +64,12 @@ variable "group_name" {
   description = "A group to control the project by being assigned group_role (defaults to project editor)"
   type        = string
   default     = ""
+}
+
+variable "create_group" {
+  type        = bool
+  description = "Whether to create the group or not"
+  default     = false
 }
 
 variable "group_role" {
@@ -179,6 +184,24 @@ variable "python_interpreter_path" {
   description = "Python interpreter path for precondition check script."
   type        = string
   default     = "python3"
+}
+
+variable "budget_amount" {
+  description = "The amount to use for a budget alert"
+  type        = number
+  default     = null
+}
+
+variable "budget_alert_pubsub_topic" {
+  description = "The name of the Cloud Pub/Sub topic where budget related messages will be published, in the form of `projects/{project_id}/topics/{topic_id}`"
+  type        = string
+  default     = null
+}
+
+variable "budget_alert_spent_percents" {
+  description = "A list of percentages of the budget to alert on when threshold is exceeded"
+  type        = list(number)
+  default     = [0.5, 0.7, 1.0]
 }
 
 variable "create_group" {