+++ outputs.tf
@@ -76,3 +76,7 @@ output "project_bucket_url" {
   description = "Project's bucket url"
 }
 
+output "budget_name" {
+  value       = module.budget.name
+  description = "The name of the budget if created"
+}